travis: Remove test-cases --race parameter
We suspect that the addition of this parameter has led to the instability
of our Travis builds over the last week or so.  Our code is simply not
yet ready to have race detection enabled on all tests.  Enabling race
detection also seems to mess up the line numbers in the stack traces,
 making it difficult to pinpoint where code is panicing and hanging.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>