ciao-image: Separate the service implementation from the data storage
service.go is the OpenStack API implementation and the data storage
implementation should be transparent to it. So they have to live in
two different packages and service includes datastore.

The next step for the datastore package is to provide a metadata
storage interface separated from the data storage one. The former
is for image metadata (noop, boltdb, etc...) and the latter is for
storing the actual image raw data.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>