ciao-launcher: Fix unit test deadlocks
The code in instance_test.go that waited for the instance loop to close
down was incorrect.  There was a possibility of deadlock if the instance
loop was sending some stats down the overseer channel at the same time
as the test was trying to shut down the instance.  There is similar code
in the overseer which actually shuts down the instance loop correctly.
This commit simply ports the good overseer code over to instance_test.go.

Partial fix for #434

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>