testutil: move information prints to stderr
There was a bug in Go testing package, which ultimately resulted in a
change to documentation instead of code.  With the (bad) Go behaviour
kept and the documentation changed, test-cases had to change how it
captures test output.  We're trying to be verbose, but show only test name
and pass/skip/fail normally, storing and outputting the verbose output only
when a test fails.  For info to be captured for later output on failure
cases, the output needs to be on stderr.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>