openstack block: extend service interface for DetachVolume
Add an endpoint handler for "os-detach".  This is a seemingly undocumented
endpoint, where "os-force_detach" is instead present.  But the
documentation for that reads like it does not do what we want (ie: "Rolls
back an unsuccessful detach operation after you disconnect the volume.").
And gophercloud does not implement it anyway.  Gophercloud does implement
the "os-detach", so treat that as the documentation and go with it.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>