testutil: goroutine server test shutdown
The ssntp shutdown will trigger all clients to send a disconnect event
with the testutil code doing a blocking write to the results channel.
But there is there is no reader of that notification unless the shutdown
is in a goroutine or the shutdown finishes.  If shutdown blocks for any
reason, the channel readers never run.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>