testutil: use consistent timeout for results channels
These timeouts "should never trigger", but if they do we get a test
failure sooner than waiting for go test's very long timeout.  The 25
second number isn't special.  And it's known to possibly be too short
for example in the case of TestReconnects(), where the ssntp client
incurs a series of delays that are 0..5, 0..10, 0..20, 0..40 seconds.
If the server is slow in restarting it _might_ take 4 seconds to restart
in travis.  But worst case we hit client delays of 1,1,1,40 seconds.
In which case despite everything "working correctly" the test case will
timeout for lack of NodeConnected result after 25s, when it would be
coming at 43s.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>