osprepare: Include an early bootstrap call for Ceph
As all the Ciao components need to have cephfs mounted before fully configuring,
and knowing the full dependency set, we have a core set of dependencies to deal
with first.

These are common to all Ciao components, and are considered the core bootstrap.
Essentially, they add dependencies on the ceph userspace tooling to ensure that
we can first mount, before we ever get to dealing with PrepareOsDeps in the
more generic sense.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>