ciao-controller: modify travis to test generation of API Documentation
API Documentation will be autogenerated every release, it will be using
"github.com/yvasiyarov/swagger" and its standards, new PRs should be
validating that they dont impact the generation of the docs, Travis
needs to verify if the docs can be generated then OK, if not the PR
should fail.

This commit contains the required changes to enable Travis to validate
it.

Brief explanation of Swagger, swagger follows a declarative comments
form to generate new docs, it will list all directories of the
component, and look for the imports, and check if the dependencies are
in the $GOPATH, vendoring does not apply, it needs the dependenceis in
order to document Models (structs, our case the payloads).

Signed-off-by: Leoswaldo Macias <lmaciasm10@gmail.com>