payloads: Add support for Storage configuration
Add a storage configuration layer to load configuration for the
SecretPath and ID for the CephDriver. This will be used in place of
having hard coded default values for the ciao-controller and
ciao-launcher in a later change.

Also add a warning for when the configuration template is missing the
storage configuration options. This is intended to become an error
later.

Signed-off-by: William Douglas <william.r.douglas@gmail.com>