ciao-controller, ciao-launcher: Use storage conf
Enable the use of the CephDriver storage configuration settings for the
ciao-launcher and ciao-controller in place of hardcoded settings.

These are currently seen as optional settings and also are able to be
overridden from commandline options. The commandline options do not
supply a default value (which in the case of the ciao-controller is a
change from previous behavior) supplied to them however, as the
configuration settings are intended to become required in the near
future.

Signed-off-by: William Douglas <william.r.douglas@gmail.com>