osprepare: allow common net and compute launcher deps
osprepare's PackageRequirements type doesn't need to use a pointer.  In
that case it's easier to compose two PackageRequirements instances which
back on a common list of type PackageRequirement.

With that, and the prior patch's addition of a Role() getter function to
the ssntp client interface, ciao-launcher can articulate common and unique
dependencies for the Agent and NetAgent role and insure the applicable set
is applied at runtime.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>