osprepare: don't add empty dependencies to the required list
It's useful for the deps.go to declare an empty need explicitly.  Ie: the
auther is stating nothing is needed on a particulat OS.  Then the deps.go
is set up like a template and a simple copy/paste/edit of a line gives
a nice oneliner diff to go from an explicit no deps to an explicit dep.
But as implemented this meant we'd try to `rpm -i ""` and emit a silly
"error" that was ignorable.  Better to just not get there and short circuit
earlier.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>