ciao-controller: confirm tenant in openstack volume interface
Prior to doing any work in the openstack_volume interface, make
sure that we have confirmed the tenant. If this is a tenant we
don't know about, we need to add them to our tenant list and
start up a cnci for them. This is permissible because they have
passed authentication already.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>