travis: Fix Go 1.6 builds
The Go 1.6 builds are currently failing as the swagger documentation checks
require the latest version of ciao's dependencies to be downloaded to the
GOPATH and built.  Sadly, some of the latest versions of our dependencies,
notably docker, no longer compile with 1.6, and so the build fails.  This
commit simply disables the swagger checks on 1.6 builds.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>