ciao-controller: rename "context" variables
The "context" variable in controller has been a local reference to the
controller state.  But now "context" is becoming a special entity in the
standard library as of golang 1.7, allowing for example cancellations,
which we will want to use.  It gets confusing if the language special name
is also used in a different way in our own code, so controller's "context"
is becoming "ctl".

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>