ciao-cli: Support specifying CA file in environment or command line
With this change the CA certificate file used for communicating with
keystone and the controller can be specified:

a.) either on the command line with -ca-file,
b.) or with an envionment variable, CIAO_CA_CERT_FILE,

thus avoiding the need to import the certificate into your local store.

This change required switching from openstack.AuthenticatedClient() to
the lower level API in order to override the transport options on the
HTTP client.

Fixes #180

Signed-off-by: Rob Bradford <robert.bradford@intel.com>