ciao-image: Add delete image support
This commit adds support for deleting images. Now, you will be able to
delete images with the command:

  ciao-cli image delete -image 78306709-9720-417e-97c6-7b941cb5565b

If the operation was successful, you will get a 204 HTTP return code.
If not, it may return the proper error, for example a 404 error could
come when the image does not exist.

Signed-off-by: Munoz, Obed N <obed.n.munoz@intel.com>