ciao-launcher: Implement boot from volume.
This commit adds support for boot from volume.  Launcher will try
to boot from volume if the ImageUUID is not present in the workload
and the workload specifies a bootable volume.

It also changes the Volumes member of vmconfig from a map to a slice.
This will be important when the start payload allows us to attach
more than one volume on boot as we will need to ensure that the
name of the devices created in the guest VM are predictable, e.g.,
the first volume is /dev/vdc, the second is /dev/vdd.

Fixes issue #367

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>