ciao-launcher: Use docker ContainerStats api to get memory and CPU usage
When there were multiple processes running inside the container the
reported memory and CPU usage values were incorrect as they only
reflected the first process created in the container.

This change makes use of the Docker ContainerStats API to obtain the
memory and CPU usage for a container.

This was tested by comparing the results from "ciaolc istats" with
"docker stats".

Fixes: #17

Signed-off-by: Rob Bradford <robert.bradford@intel.com>