ciao-launcher: Remove now unnecessary PID code
The PID of the starting process in the container was obtained
and stored for the purpose of retrieving the memory and CPU usage of the
container. As we now get that information from Docker via the
ContainerStats API it is no longer necessary to query the PID and store
it.

This is cleanup following the work to fix #17

Signed-off-by: Rob Bradford <robert.bradford@intel.com>