Bring ssntp/examples up to date with latest API
Examples in ssntp needed to be updated to work properly with the
latest version of the ssntp package. This patch brings the examples up
to date on the naming of the certificates from ciao_ca_cert to ca_cert
and usage of the Role struct needing to be changed to ssntp.Role.

Signed-off-by: John Andersen <john.s.andersen@intel.com>