test-cases: Fix package list
test-cases contained some uncessary code for handling the list of packages
passed to it on the command line.  This code was unnecessary as go list
already does this processing for us, and it does it better.  So this patch
simply removes the custom handling, and test-cases now behaves like all
the other go tools, which is what users expect.

The commit also updates the usage message to explain how to specify packages
on the test-cases command line.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>