osprepare: support context.Context for cancellation
The golang package "context" allows cancellation to flow across API
boundaries and processes.  Launcher spawns all manner of processes so
it takes extra care to allow for example clean exiting of both itself
and its children through a control-C of running launcher.

Since osprepare may do a long-running operation in the (ideally rare) case
that launcher dependencies are not present at start up.  The user may want
to break out of this and not let it run to completion and not actually
start launcher.  That is only possible if a context is passed.

For controller and scheduler I simply use the context.TODO() as the package
documentation says not to pass a nil context and rather one should use this
marker empty context.

Fixes https://github.com/01org/ciao/issues/500

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>