ciao-launcher: Add volume support to docker containers
This commit add the ability to automatically mount a ceph volume
into a docker container when that container is created.  The location
of the mount point inside the container is currently set to
/volumes/UUID.  This will be configurable in the future.  All ceph
volumes are currently mounted RW.  Again this is something that will
be congifurable in the future.  Volumes tagged as bootable cannot
be mounted inside a docker container and any workload containing
such volumes will fail to start.

Fixes #527

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>