osprepare: avoid duplicate InstallDeps()s in a component
For an entity like launcher which can run in two modes, each of which may
have distinct dependencies, we want to run osprepare once on the union of
the dependencies.

This patch starts extending osprepare to have helpers for managing
manipulations of PackageRequirements.  First is an Append() helper,
and later will be an intelligient deduplication (eg: picking the highest
version number of overlapping packages).

Launcher is modified to call osprepare.InstallDeps() once on such a merged
list of dependencies.

Fixes https://github.com/01org/ciao/issues/572

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>