ciao-controller: begin fixing getStorage
getStorage needs to handle a lot more cases than it currently does.
Implement the case where we need to create a volume from an image
or when the user wants to create a volume from another volume,
or an empty volume when the user requests a storage resource in a
workload template.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>