ciao-launcher: Map and mount docker volumes on start
This commit delays the mapping and mounting of docker volumes to just
before the container starts.  Previously, we were performing these steps
earlier, namely before the image was created.  We were also unmapping and
unmounting when workloads are deleted.  This led to ceph volumes being
mapped when they were not actually being used and also to missing volumes
when restarting containers after a reboot.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>