Single VM: Download Fedora images in setup.sh
The Single VM defines a workload for a Fedora image but it does not
provision that image in to the /var/lib/ciao/images folder.  This
means that users cannot create a Fedora instance even though a
Fedora workload appears to exist when running ciao-cli workload list.

This also causes the BAT tests to fail inside single VM as these
tests randomly start one of the predefined workloads.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>