documentation: enable swagger to use vendoring
Swagger is a tool used to generate documentation based on
declarative comments in the code, the problem with it was that it
had an issue with vendoring. We have fixed that issue in
https://github.com/yvasiyarov/swagger/pull/120
and now we can use our vendor directory. This fixed was need because
we were bringing dependencies with 'go get' which in some PRs were
causing to fail for time outs, an issue aside of the PRs.

In this change we add the required argument to enforce to check for
vendor directory and remove the 'go get'.

Fixes #559

Signed-off-by: Leoswaldo Macias <leoswaldo.macias@intel.com>