Single VM: Let launcher clean up containers
The setup.sh and run_launcher.sh scripts executed commands to delete
all the docker containers present on the Single VM.  There's no need
to do this as the invocation of launcher with the --hard-reset flag
will clean up any containers created by ciao.  It's also a little
destructive as the old scripts will destroy any non-ciao related
containers, which is probably not what we want.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>