Travis Cluster Test Automation
A new flag -t has been added to identify a travis run vs a normal
run. This allows the test harness to do any travis specific
setup, and handle latency issues due to the fact that travis
VM's are slower than bare metal and normal VMs.

Leverage the single machine test setup to verify a virtual cluster
on travis. Do not test VM's in travis for now.

The test will
- Provision a virtual cluster
- Launch two containers
- Check network connectivity between containers
- Tear down the cluster

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>