ciao-cli: support creation of bootable volumes
In order to create a bootable volume, you must specify an
image UUID. Since you can also create a volume from another volume,
add a "source_type" option which can be either "image" or "volume"
and fill in either SourceVolID or ImageRef based on the value
of this type. In ciao-controller, if ImageRef is not nil, we'll
create a volume from an image. If SourceVolID is not nil, we'll
copy an existing volume.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>