Single VM: Move tests into verify.sh
This commit moves the tests that the Single VM runs to verify that
everything is setup correctly to a separate file called verify.sh.
The reason is that if you are testing out a small change on the
Single VM, you don't want to have to re-run these tests everytime
you run setup.sh, as doing so can take several minutes.

Fixes #536

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>