Single VM: Tidy up setup.sh's output
This commit cleans up the output from the setup.sh script.  Previously,
it was outputting instructions from the controller logs which gave the
impression that the user needs to follow these instructions, which he/she
doesn't.

A new line and a horizontal bar were also added to separate the script's
logs and the instructions that the user does need to follow.

Signed-off-by: Mark D Ryan <mark.d.ryan@intel.com>