ciao-cli: Make volume and image service work with -ca-file
The volume and image service were not passing the CA certificate,
specified either on the command line or via an environment
variable, down to gophercloud.  This meant that these services
could not work inside the Single VM.  As we now have three different
gophercloud services requiring the same authentication code, the
original code that implemented this feature for the identity service,
has been moved to a new function.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>