ciao-api: implement http api call to assing floating IP
Floating ips are assigned to the CNCI agent and the it makes
the forward of traffic to the instances. In this commit
we implement the api call to assign floating APIs.

We are based on http://developer.openstack.org/api-ref-compute-v2.1.html#addFloatingIp.
The api call is /v2.1/{tenant_id}/servers/{server_id}/action using
POST, and sending the parameter 'addFloatingIp'

Signed-off-by: Leoswaldo Macias <leoswaldo.macias@intel.com>