Networking: Make the boltDB+gob table database generic
Make the boldDB+gob table database generic so that it can
be reused by other components.

Note: As the table name internally has changed, this will break
any in place update of the CN launcher. For a proper update
of the CN, it has to be hard reset to remove all instances
and the database cleaned up.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>