test-cases, travis: Add support for TAP reports
This commit adds test-cases support for the generation of TAP reports. It also takes
the oppurtunity to rationalize the options for specifying the output format.  There
is now one format option, --format, which accepts a number of different values;
html, tap, colour-text and text.  The default is colour-text.

This change requires a modification to the travis script that executes test-cases,
as the old formatting option used by travis.yml no longer exists.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>