openstack: Don't close request bodies in handlers
The image service code seems to be closing the bodies of received requests in
its handlers.  This breaks in Go 1.8 as these bodies are likely to be nil.
Also, unless I'm mistaken, there's no need for the server to close the bodies.
This is only needed for clients that pass a file or a socket to http.NewRequest.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>