ciao-controller: Fix volume detail "corruption" in ListVolumesDetail
Various members of the Volume structure require a pointer. These were
being filled by taking the address of the member from the loop variable.
Unfortunately because this loop variable is reused for each iteration of
the loop all these pointers were ultimately the same resulting in
duplicated values in the volume details list.

The solution is to loop through the slice by index and take pointers to
the underlying data.

Fixes: #644

Signed-off-by: Rob Bradford <robert.bradford@intel.com>