ciao-cli: Fix instance list template ordering
The commits that sorted the result of the instance list command and
added template support for this command conflicted.  The automerge
done by github was insufficient and resulted in an unsorted slice
being passed to the user supplied template.  This commit simply
re-arranges the code a little bit to ensure that the instances processed
by the user template are sorted.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>