ciao-launcher: Fix boot from volume
Boot from volume only worked properly when launcher was run with the
--with-ui spice option.  Without this option, boot from volume instances
start, appear to be active, but don't actually boot.  The problem
was caused by the use of an incorrect PCI address for the rootfs.  The
PCI address for the rootfs is influenced by the --with-ui parameter
and launcher was not taking this fact into account.

Fixes: #751

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>