ciao-launcher: Fix docker network locks
This commit fixes #763, an issue that was introduced by some poor optimization.
When we create a container on a subnet on a CN for the first time we need
to create a new docker network.  This network needs to be created after the vnic
but before the container is created.  It follows that we need to be careful when
creating multiple containers in parallel.  The docker network will be created
during the creation of one of these containers.  The instance go routines
controlling the creation of the other containers need to block until the network
has been created, otherwise they will fail.  The optimisation deleted by this PR
was trying to reduce the amount of time the lock would be held for.
Unfortunately, the code did not consider the case that when instances are
deleted, the final instance deleted for a subnet on a CN might be a VM and not a
container.  When this happens, it becomes impossible to start a new container
for that subnet as the container code thinks that the network for that subnet
already exists, when in fact the underlying bridge was deleted along with the
VM.

This commit fixes the problem by removing the optimisation and placing Vnic
and docker network creation inside a critical section.  As we do not currently
allow parallel vnic creation, this shoudln't actually slow anything down.

Fixes #763

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>