ciao-controller: Boot from an image volumes using copy-on-write
If a volume with the same ID as the image exists then a new volume will
be created as a copy-on-write clone of that existing image volume.

If it does not and the file can be found on the local disk the image
will be uploaded into a new volume with the same ID as the image. A
copy-on-write clone is then created from that newly created volume.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>