ciao-launcher: Stop leaking docker networks
There were two issues here.  Commit 54d8e50 introduced a bug which
lead to a new docker network being created with every container,
which is not what we want.  The intended behaviour is to create a
new docker network for each new subnet.  This creation of extra
networks lead to a leak as they were not getting deleted as they
were not supposed to exist.

There was an existing bug which could result in docker network
leaks when you created a container, a VM, deleted the container
and then the VM in that order.

Both these issues have now been fixed.

Fixes #768

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>