ciao-networking: simplify cnci image preparation mount
We have a race potentially between losetup and partition existance.  This
probably is what the retry loop managed to kind of work around in the past.
The udevadm settle was in the wrong location previously.  It needs to come
after creation and before use.  It would have coincidentally in the past,
IF you were retrying.

We can simplify by not looping and rather create, wait, use as a simple
series of commands.

If we find subsequent issues that need more logic in this flow, I suggest
we remove the losetup and udevadm settle, and instead do "kpartx -sav
$img" followed by the mount.  Then also on exit do "kpartx -sdv $img".

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>