ciao-controller: Store correct tenant ID for instance deleted events
The controller was incorrectly storing the instance ID in the tenant ID
field of the log table for instance deleted events.  This meant that it
was not possible for the user to view these events using ciao-cli event
list as events are filtered by tenant ID, and the IDs stored did not
correspond to any tenants.

Fixes #776

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>