ciao-image: Fix issue sending image ID on Image creation
This commit fixes an issue that occured when an ID was passed on
Image creation request, it was return an HTTP 409 error code in all
requests. Now, it sends a 201 (Success HTTP Code) when the image
does not exist, and it will return the proper error when the ID
exists or there was an internal db error.

Signed-off-by: Munoz, Obed N <obed.n.munoz@intel.com>