contributing: Let the style sheet through for verbatim data
One of the commit messages was using 8 spaces indentation, the other one
was using tabs. Github md renderer/style sheet was using 4 spaces to
represent a tab in verbatim data.

The result was that the two commit messages weren't properly aligned.

Changed it to remove manual indentation and let the renderer/style sheet
decide.

An alternative could be to always indent with the same character(s), say
4 spaces so the md "source" looks good as well.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>