ansible: Optionally skip ceph configuration on ciao nodes
Ciao requires a ceph cluster to be available for all ciao nodes.

The deployment playbook requires that the systadmin provides ceph
configuration and secret files in order to configure ceph on all
ciao nodes.

This changes introduces `skip_ceph` variable which allows the deployment
to skip ceph configuration and leave that step to be performed manually
by the sysadmin.

Fixes #782

Signed-off-by: Alberto Murillo Silva <alberto.murillo.silva@intel.com>