testutil: Add snake oil cert files for roles
The current test utility support passing role based snake oil
certificates for testing but only only as strings. Some of the APIs that
need to be tested require these to be passed as file paths instead so
this patch adds that capability.

The fake ca certs were generated with the following process using
ciao-cert from commit 3bb33ad64383f0712d0f68e7e25379a7be636254:

$ ./ciao-cert -directory ../testutil -host localhost -server -role \
scheduler -ip 127.0.0.1

$ ./ciao-cert -directory ../testutil -host localhost -role $role \
-server-cert ../testutil/cert-Scheduler-localhost.pem -ip 127.0.0.1

The RoleToTestCertPath function the change implements will give the path
to the CA certificate generated when creating the server certificate and
the path to the certificate matching the requested role. If there is no
certificate path matching the specified role an error will be returned
as well.