ciao-launcher: Send InstanceDeleted after instance deletion
Launcher was incorrectly sending the InstanceDeleted event before the
instance and its associated network assets had been deleted.  This
was causing a race condition in which controller was re-using some
of the resources assigned to the instance before it was actually deleted for
new instances.  When this happened the new instances typically failed to
start.  This commit delays the sending of the InstanceDelete event until
after the instance has been deleted.

Fixes #808

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>