Single VM: Cleanup any orphan ciao networks
When Single VM tests fail and leave orphan network behind, even
after fixing the issue the tests will not run successfully.

This is due to the fact that any orphan ciao network interfaces
will not exist in the ciao database, but will continue to exist
in the docker database.

Cleanup any orphan ciao networks so that subsequent runs do not fail
due to old orphan networks.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>