ciao-controller: let volume requests flow to startWorkload()
This commit introduces plumbing to allow a generic volume list argument in
the ciao-controller's startWorkload() command, but does nothing with it.
This allows us to easily see we've accounted for existing call sites
in test code and the code launching a tenant CNCI instance, which shows
current code continues to compile and run correctly.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>