ciao-cli: Improve template functionality
This PR adds two new functions to the template language understood
by ciao-cli:

- tojson outputs the specified object in json format, e.g., {{tojson .}}
- filter operates on an slice or array of structures.  It allows the caller
  to filter the input array based on the value of a single field.
  The function returns a slice containing only the objects that satisfy the
  filter, e.g.

  ciao-cli image list -f '{{$x := filter . "Protected" "true"}}{{len $x}}'

  outputs the number of protected images maintained by the image service.
- select operates on a slice of structs.  It outputs the value of a specified
  field for each struct on a new line , e.g.,
  {{select . "Name"}}

The image list command was also modified to called the template utility
functions in template.go.  This was necessary to ensure that our new
functions are available to the image list command.

The help documentation for commands that accept templates has been updated
to provide a descriptions for the new functions.

Finally, a the createTemplate function was using the wrong name for the
templates it created.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>