singlevm: Add hosts entry in keystone container
Feeds the host's IP address and hostname to docker when starting the
keystone container, so that it will add the appropriate entry to
the container's /etc/hosts. This allows the openstack commands within
the container's bootstrap.sh to run (passing certificate validation).

Adds "vlan_" into the names of the prior ciao_ip variable and its peers
to disambiguate.

Signed-off-by: Brett T. Warden <brett.t.warden@intel.com>