ciao-controller: Add the ability to switch DB backend for datastore
The datastore has gained a new member, DBBackend which is a
persistentStore interface. Using this the backend for the datastore can
be specified. If no backend is specified then then the sqlite backend is
used.

To ensure this functionality works the datastore test was updated to use
this new member.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>