ciao-cert: Rename signer flag for clarity
Previously the -server and -server-cert flag were used to indicate that
the certificate being generated would be either the signing certificate
or signed by the signing certificate. This change updates the flags to
be -anchor and -anchor-cert to separate the ssntp.SERVER role term from
the trust anchor term used in the same command and APIs.