ciao-controller: add volumes to newInstance/newConfig
The controller's startWorkload creates an instance with a config,
where the config is articulated based on the struct payloads.Start,
which includes an array of payloads StorageResource type objects.
Each storage.BlockDevice in the array of instance volumes coming into
startWorkload() for instance creation needs mapped to a similar object in
the Start yaml for the instance config.  This is slightly complicated
by controller having a 3rd similar data types.StorageResources and
the need for controller to track existing volumes in the similar-ish
BlockData struct and volume-to-instance mappings with the similar-ish
StorageAttachments struct.

For the most part, existing code passes an empty volumes array into
newConfig.  From the compute API, a startWorkload() -> newInstance() ->
newConfig() should result in a set of StorageResources in the Start yaml
payload, with some variations including having the controller create the
volume.  Others will have ciao-launcher create a volume.  In either case,
ciao-launcher would take the appropriate steps to attach the volume to the
instance of the workload, whether VM, container or bare metal.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>