ciao-controller: deduplicate code
There are three sets of nearly identical code in getStorage which can
be abstracted.  The primary difference between them was whether or not
they care about the storage being ephemeral or bootable.  We actually
already have data members in the ciao-controller/types/types.go
StorageResource structe for Bootable and Persistent, so use them.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>