ciao-controller: move getStorage to use simpler arg
When a workload can have multiple volumes declared (as per issue #694)
the workload will have an array instead of a pointer and getStorage()
will be called for each StorageResource object in the array.

This leaves the previously inserted addBlockDevice() helper more readily
able to also deal with an individual StorageResource, instead of requiring
a pointer input.

My default rule of thumb in golang is to not use a pointer unless I have
a specific reason to, and normally simply trust in the memory manager.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>