ciao-storage: extend interface to include snapshot uuid validator
The form of a snapshot UUID and its associated input validation and usage
will vary with storage backend type.  Today there's just ceph, for which
Ciao's snapshots will be of the form {volumeUUID}@{snapshotUUID}.  But
rather than code checks based on a helper that uses that specific form,
make it abstract so future backends are easier to add.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>