singlevm: remove admin from "service" project
The keystone container's bootstrap gives the admin user an admin role in
the service project. Later, we give the admin user an admin role in the
admin project, which is more appropriate. Having roles in multiple
projects complicates the use of admin to manage tenants, so we remove
the extraneous role in the service project.

Signed-off-by: Brett T. Warden <brett.t.warden@intel.com>