ciao-controller: Replace sleeps with channel signalling
The wrapper implements the same interface as the client at the core of
controller. This wrapper proxies all commands and events to the real
client. Users of the wrapper can receive notification whenever the
underlying client has completed execution of an Event or Error handler.

This allows the removal of the sleeps from controller_test.go since they
can now be replaced by the use of channel signalling from the wrapper.

Fixes: #83

Signed-off-by: Rob Bradford <robert.bradford@intel.com>